Add rotational field to ocf_core struct
This property is necessary if we don't want to sort cache lines
or flush containers when the backend device is not an HDD.
Sorting cache lines or flush containers on non-HDDs is a waste of time.
Add functions to work with rotational property.
Make the cleaner's cache line sorting dependent on the core volume type.
If any of the core volumes is HDD, sort cache lines, otherwise do not.

Signed-off-by: Daniel Madej <daniel.madej@huawei.com>